[shrks-55] Some fix
